ended to split resolvers and schemas
