+++ src/schemasResolvers/schema.ts
@@ -2,9 +2,9 @@ import { makeExecutableSchema } from "apollo-server";
 import {merge} from "lodash";
 
 import {typeDef as Assets, resolvers as assetsResolvers} from "./assets"
-// import {typeDef as Auth} from "./assets"
-// import {typeDef as User} from "./assets"
-// import {typeDef as Workspace} from "./assets"
+import {typeDef as Auth, resolvers as authResolvers} from "./auth"
+import {typeDef as User, resolvers as userResolvers} from "./user"
+import {typeDef as Workspace, resolvers as workspaceResolvers} from "./workspace"
 
 const Query = `
 type Query {
@@ -19,7 +19,6 @@ type Mutation {
 // const resolvers = {}
 
 export const schema = makeExecutableSchema({
-  typeDefs: [Query, Mutation, Assets],
-  resolvers: merge(assetsResolvers)
-//   typeDefs: [Assets, Auth, User, Workspace]
+  typeDefs: [Query, Mutation, Assets, Auth, User, Workspace],
+  resolvers: merge(assetsResolvers, authResolvers, userResolvers, workspaceResolvers)
 })
